+++ src/Contracts/Routable.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Astrotomic\Stancy\Contracts;
+
+interface Routable
+{
+    public function getUrl(): string;
+}