+++ src/StancyServiceProvider.php
@@ -2,38 +2,39 @@
 
 namespace Astrotomic\Stancy;
 
+use Astrotomic\Stancy\Contracts\ExportFactory as ExportFactoryContract;
 use Astrotomic\Stancy\Contracts\FeedFactory as FeedFactoryContract;
 use Astrotomic\Stancy\Contracts\Page as PageContract;
 use Astrotomic\Stancy\Contracts\PageFactory as PageFactoryContract;
 use Astrotomic\Stancy\Contracts\SitemapFactory as SitemapFactoryContract;
+use Astrotomic\Stancy\Factories\ExportFactory;
 use Astrotomic\Stancy\Factories\FeedFactory;
 use Astrotomic\Stancy\Factories\PageFactory;
 use Astrotomic\Stancy\Factories\SitemapFactory;
 use Astrotomic\Stancy\Models\Page;
+use Illuminate\Contracts\Support\DeferrableProvider;
 use Illuminate\Support\ServiceProvider;
 
-class StancyServiceProvider extends ServiceProvider
+class StancyServiceProvider extends ServiceProvider implements DeferrableProvider
 {
-    public function boot(): void
-    {
-        $this->publishes([
-            __DIR__.'/../config/stancy.php' => config_path('stancy.php'),
-        ], 'config');
-    }
-
     public function register(): void
     {
-        $this->registerConfig();
         $this->registerPage();
         $this->registerFeed();
         $this->registerSitemap();
+        $this->registerExporter();
     }
 
-    protected function registerConfig(): void
+    /** @codeCoverageIgnore */
+    public function provides(): array
     {
-        $this->mergeConfigFrom(
-            __DIR__.'/../config/stancy.php', 'stancy'
-        );
+        return [
+            PageFactoryContract::class,
+            PageContract::class,
+            FeedFactoryContract::class,
+            ExportFactoryContract::class,
+            SitemapFactoryContract::class,
+        ];
     }
 
     protected function registerPage(): void
@@ -50,5 +51,10 @@ class StancyServiceProvider extends ServiceProvider
     protected function registerSitemap(): void
     {
         $this->app->singleton(SitemapFactoryContract::class, SitemapFactory::class);
+    }
+
+    protected function registerExporter(): void
+    {
+        $this->app->singleton(ExportFactoryContract::class, ExportFactory::class);
     }
 }