+++ src/Traits/PageHasUrl.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Astrotomic\Stancy\Traits;
+
+use Spatie\Sitemap\Tags\Tag;
+use Spatie\Sitemap\Tags\Url;
+
+trait PageHasUrl
+{
+    abstract public function getUrl(): string;
+
+    public function toSitemapItem(): Tag
+    {
+        return Url::create($this->getUrl());
+    }
+}