+++ tests/TestCase.php
@@ -2,6 +2,7 @@
 
 namespace Astrotomic\Stancy\Tests;
 
+use Astrotomic\Stancy\Contracts\ExportFactory as ExportFactoryContract;
 use Astrotomic\Stancy\Contracts\FeedFactory as FeedFactoryContract;
 use Astrotomic\Stancy\Contracts\PageFactory as PageFactoryContract;
 use Astrotomic\Stancy\Contracts\SitemapFactory as SitemapFactoryContract;
@@ -9,6 +10,7 @@ use Astrotomic\Stancy\StancyServiceProvider;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\View;
 use Orchestra\Testbench\TestCase as Orchestra;
+use Spatie\Export\ExportServiceProvider;
 use Spatie\Feed\FeedServiceProvider;
 use Spatie\Sheets\SheetsServiceProvider;
 use Spatie\Sitemap\SitemapServiceProvider;
@@ -26,17 +28,14 @@ abstract class TestCase extends Orchestra
     protected function getPackageProviders($app): array
     {
         return [
+            // dependencies
+            ExportServiceProvider::class,
+            FeedServiceProvider::class,
+            SheetsServiceProvider::class,
+            SitemapServiceProvider::class,
+            // stancy
             StancyServiceProvider::class,
         ];
-    }
-
-    protected function getApplicationProviders($app): array
-    {
-        return array_merge(parent::getApplicationProviders($app), [
-            SheetsServiceProvider::class,
-            FeedServiceProvider::class,
-            SitemapServiceProvider::class,
-        ]);
     }
 
     protected function getEnvironmentSetUp($app): void
@@ -49,11 +48,29 @@ abstract class TestCase extends Orchestra
             'driver' => 'local',
             'root' => realpath(__DIR__.'/resources/content/blog'),
         ]);
+        $app['config']->set('filesystems.disks.export', [
+            'driver' => 'local',
+            'root' => realpath(__DIR__.'/export'),
+        ]);
+        @mkdir(__DIR__.'/export', 0777);
 
         $app['config']->set('sheets.collections', [
             'content',
             'blog',
         ]);
+
+        $app['config']->set('feed.feeds', []);
+        $app['config']->set('feed.feeds.blog', [
+            'title' => 'test blog feed',
+            'items' => [FeedFactoryContract::class.'@makeFromSheetCollectionName', 'blog'],
+            'url' => '/feed/blog.atom',
+            'view' => 'feed::atom',
+        ]);
+
+        $app['config']->set('export.disk', 'export');
+        $app['config']->set('export.crawl', false);
+        $app['config']->set('export.clean_before_export', true);
+        $app['config']->set('export.include_files', []);
     }
 
     protected function getRequest(string $url = '/', array $headers = []): Request
@@ -83,5 +100,10 @@ abstract class TestCase extends Orchestra
     protected function getSitemapFactory(): SitemapFactoryContract
     {
         return $this->app->make(SitemapFactoryContract::class);
+    }
+
+    protected function getExportFactory(): ExportFactoryContract
+    {
+        return $this->app->make(ExportFactoryContract::class);
     }
 }